Bump fs2 version
